Added routing and a page for skyline
