+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './pages/App';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
 import '@fontsource/inter';
