+++ src/pages/SkylinePage.tsx
@@ -1,0 +1,11 @@
+import SingleFoldPageUIWrapper from "../components/SingleFoldPageUIWrapper";
+
+function SkylinePage() {
+	return (
+		<SingleFoldPageUIWrapper>
+			Hello!
+		</SingleFoldPageUIWrapper>
+	)
+}
+
+export default SkylinePage;
